feat(theme): apply the color theme app-wide (all tabs), not just EDIT
Moves the .edit-theme-scope + inline --et-* vars from the WaveformEditor root up to the Shell root, so the selected theme recolors every tab and the app chrome. The EDIT root reverts to its original bg-black/40 (now themed by the global surface remap). Default --et-root-bg becomes #07050a (the Shell canvas) so Midnight is pixel-identical; light-theme text overrides broadened for app-wide legibility. Modals portaled to <body> stay outside the scope (default dark chrome).
